requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.0.3
gspread==5.10.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1